requirements.txt
python-magic==0.4.15
python-magic-bin==0.4.14
termcolor==1.1.0